+++ lib/ruby_aem/resources/config_property.rb
@@ -44,11 +44,12 @@ module RubyAem
       def create(run_mode, config_node_name)
 
         name = RubyAem::Swagger.property_to_parameter(@call_params[:name])
+        type_hint_prefix = name.gsub(/^_/, '')
 
         @call_params[:run_mode] = run_mode
         @call_params[:config_node_name] = config_node_name
         @call_params["#{name}".to_sym] = @call_params[:value]
-        @call_params["#{name}_type_hint".to_sym] = @call_params[:type]
+        @call_params["#{type_hint_prefix}_type_hint".to_sym] = @call_params[:type]
 
         config_name = Swagger.config_node_name_to_config_name(config_node_name)
         @client.call(self.class, __callee__.to_s.concat(config_name.downcase.gsub(/\s+/, '')), @call_params)