+++ test/unit/resources/config_property_spec.rb
@@ -26,6 +26,21 @@ describe 'ConfigProperty' do
       config_property.create('author', 'org.apache.felix.http')
     end
 
+    it 'should call client with expected parameters when property is a keyword' do
+      expect(@mock_client).to receive(:call).once().with(
+        RubyAem::Resources::ConfigProperty,
+        'createapacheslingdavexservlet',
+        { :name => 'alias',
+          :config_node_name => 'org.apache.sling.jcr.davex.impl.servlets.SlingDavExServlet',
+          :type => 'String',
+          :value => '/crx/server',
+          :run_mode => 'author',
+          :_alias => '/crx/server',
+          :alias_type_hint => 'String' })
+      config_property = RubyAem::Resources::ConfigProperty.new(@mock_client, 'alias', 'String', '/crx/server')
+      config_property.create('author', 'org.apache.sling.jcr.davex.impl.servlets.SlingDavExServlet')
+    end
+
   end
 
 end